make oop code for kafka folder
